separate deep_sleep component for each platform in different file
